Update Versão_não_única_2.0.c
Agora sim, tinha trocado pra float sem querer

+++ C/Versão_não_única_2.0.c
@@ -80,7 +80,7 @@ void soma(int vet[], int Nelementos) {
     double soma = 0;
 
     for (i = 0; i < Nelementos; i++) {
-        soma += (float) vet[i];
+        soma += (double) vet[i];
     }
 
     printf("\n[Soma total  : %.0f]\n\n", soma);
